Add files via upload
A pool of volume V has two pipes from which it is filled. Each pipe has a certain flow rate (the litres of water passing through one pipe in one hour). The worker starts the pipes simultaneously and leaves in N hours. Write a program that retrieves the results from the pool the moment the worker returns.